feat: Add question type functionality
This commit adds the functionality to specify question types when
creating and storing questions. This includes:

- Adding a `question_type_id` field to the database. - Updating the
`StoreQuestionRequest` and `StoreBatchQuestionsRequest` to include
`question_type_id`. - Adding `question_types` to the index view. -
Adding a question type select to the store and batch views. - Added
javascript to handle default question creation mode.